Fixup after merge conflicts
